insights/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
darts>=0.28.0
lightgbm>=4.3.0
holidays>=0.40
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0